+++ uefi-raw/CHANGELOG.md
@@ -4,6 +4,8 @@
 - Added `SimpleTextInputExProtocol`.
 
 ## Changed
+- Corrected the type of the `driver_image` parameter in
+  `BootServices::connect_controller` from `Handle` to `*const Handle`.
 
 
 # uefi-raw - v0.14.0 (2026-03-22)